+++ HISTORY.md
@@ -1,7 +1,10 @@
 # Rocksdb Change Log
 ## Unreleased
 ### New Features
-* Added new ChecksumType kXXH3 which is faster than kCRC32c on almost all x86_64 hardware.
+* Added new ChecksumType kXXH3 which is faster than kCRC32c on almost all x86\_64 hardware.
+
+### Bug Fixes
+* Prevent a `CompactRange()` with `CompactRangeOptions::change_level == true` from possibly causing corruption to the LSM state (overlapping files within a level) when run in parallel with another manual compaction. Note that setting `force_consistency_checks == true` (the default) would cause the DB to enter read-only mode in this scenario and return `Status::Corruption`, rather than committing any corruption.
 
 ## 6.26.0 (2021-10-20)
 ### Bug Fixes